chore(test): rename module util to utils
