--- test/src/util/mod.rs
@@ -1,13 +1,0 @@
-pub mod chain;
-pub mod check;
-pub mod message;
-pub mod mine;
-pub mod sugar;
-
-pub use chain::{download_main_blocks, download_main_headers, submit_blocks};
-pub use check::{
-    is_transaction_committed, is_transaction_pending, is_transaction_proposed,
-    is_transaction_unknown,
-};
-pub use mine::{mine, mine_until, mine_until_with, mine_with};
-pub use sugar::{out_bootstrap_period, out_ibd_mode};
