+++ test/src/utils/mod.rs
@@ -1,10 +1,8 @@
 use crate::{Net, Node, TXOSet};
-use ckb_jsonrpc_types::{BlockTemplate, TransactionWithStatus, TxStatus};
+use ckb_jsonrpc_types::{TransactionWithStatus, TxStatus};
 use ckb_types::{
     core::{BlockNumber, BlockView, EpochNumber, TransactionView},
-    h256, packed,
-    prelude::*,
-    H256,
+    h256, H256,
 };
 use std::convert::Into;
 use std::env;
@@ -12,6 +10,12 @@ use std::fs::read_to_string;
 use std::path::PathBuf;
 use std::thread;
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+
+pub mod chain;
+pub mod check;
+pub mod message;
+pub mod mine;
+pub mod sugar;
 
 pub const FLAG_SINCE_RELATIVE: u64 =
     0b1000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000;
